requirements.txt
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
python-dateutil==2.8.1
psycopg2==2.8.5
urllib3==1.25.9
#pytz==2019.3
